+++ app/Commands/Analyse.php
@@ -14,7 +14,7 @@ use Symfony\Component\Console\Helper\TableSeparator;
 
 class Analyse extends Command
 {
-    public const VERSION = '1.0.4';
+    public const VERSION = '1.0.5';
 
     private PackageAnalyser $packageAnalyser;